setup: exchange authorization code for access token
Once the authorization code is extracted from the callback, we continue the OAuth flow and exchange it for the access token that will be used as a bearer token for all Monica API calls. These calls are made using Retrofit and Moshi.

Some basic error handling has been implemented, such as in the case where the client secret is incorrect and the token API call returns an error (HTTP 500). For now, the error is a simple true/false result which is mapped to a generic error on the UI layer.

The auth and token information has been split away from the OAuthSettings so that they can be cleared separately but also because they are derived from the user-provided configuration.