dashboard: load user's avatar
Added the Photos API, which is used to fetch contact photos. The photos are stored in-memory in the Photos repository. A usecase called GetContactAvatar has been added to encapsulate the logic of figuring out a contact's avatar photo given there is no explicit link available through the API. The dashboard screen uses this usecase to load the current user's avatar by passing a ByteBuffer to Coil.