use room as source of truth for user and photo databases
These changes make Room the backing data source for both the User table (the "me" entity) and the Photos table. Data is now fetched from these databases instead of a Flow in memory.

A nice improvement that Room gives us can be seen in the MeFullDetails class. With one type, Room can join multiple tables and get all the info we need to display the dashboard properly: the user's name, initials, and photo if available. The removes the need for manually looking up multiple repositories so the GetContactAvatar approach has been removed.