home: create account bottom sheet
The account screen will hold actions that are less frequently used, such as settings or signing out.

Used compose-destinations to navigate to the new account screen and wrapped the top level DestinationNavHost in a custom ModalBottomSheetLayout with a transparent background. This will allow individual bottom sheets to have different layouts.

The material dependency was needed because of the ModalBottomSheetState class being used by accompanist.

The navigation to the account bottom sheet is done from the dashboard page which is in the nested dashboard nav graph. The account destination needs to be in the top-level nav graph so that it draws over all the content in the app. To communicate this navigation, the simplest option was to create the HomeNavigationManager which allows top-level navigation from anywhere in the app.