dashboard: add recent contacts section
The recent contacts are currently displayed in no particular order.

The logic that loads contact pictures has been changed and it is now done by a Coil Fetcher, `UserAvatarFetcher`. `UserAvatar` instances can now be provided directly to Coil as model and this fetcher will look up the Room database and load the user's avatar.