dashboard: sort recent contacts by updated date
The recent contacts should be sorted by last access date but that's not something provided by the API so the updated date is used as a stopgap.

The updated date is parsed as a `ZonedDateTime` from the API and saved to Room with the same type. Separate converters were added as needed. `ZonedDateTime` is a new API that's not available on all Android versions so core library desugaring was added.