photos: fetch all photos on sync
Instead of fetching photos individually for each contact, fetch all at once in order to reduce the risk of hitting the rate limit. This may need to be revisited in order to strike a good balance between the frequency of API calls and the amount of data loaded for the first time.