journal: add initial refresh support and paging
Due to the potentially large number of journal entries, the Paging library is introduced to avoid holding too many entries in memory.

The `JournalPagingSource` is responsible for grabbing the correct pages of data from Room. It also observes the refresh status and invalidates itself after a refresh event.

The synchronization is done on the first load of the journal screen and then only on demand.