contacts: add pagination
Apply the same paging strategy used for the journal screen on the contacts screen.

There seems to be a bug with the response returned from Monica where multiple contacts can have the exact same updated_at date down to the second. This breaks pagination because, when scrolling, it's possible to bind the same contact to multiple positions depending on how the results are returned from the database. To fix this for now, sort the results on the contact screen alphabetically.