feat(activities): add initial activity creation screen
To support navigating to the new screen, the DestinationsNavigator is now passed through to all the contact detail sections.

The activities FAB should be present for both empty and loaded states so the Scaffold was moved up.

The bottom bar should not be taking up screen space when viewing a contact or editing an activity so an animation has been added to display it only on the 3 home tabs.