feat(journal): redesign entry UI
The entry UI is now cleaner and it includes the date and moves the save action to a FAB.

Additionally, split up the `JournalEntryUiState` into separate UI classes, one for the list screen and one for the entry screen.

Fixed the fact that a UI class was part of the Journal repository. Make the transformation from domain to UI class at the ViewModel level instead.