feat(gradle): add build-logic
In order to support more modules in the project, it's currently good practice to have a `build-logic` folder that holds common configuration logic between module. This was inspired by NowInAndroid which in turn was inspired by Square.

This adds the `build-logic` module along with an Android application plugin which encapsulates the logic around it.