feat(configuration): support clearing DB on next launch
Added a new Configuration activity meant to control parameters outside of general app settings. It is available to be launched through a shortcut at any time.

For now, this allows clearing the Room database on the next app launch. This is helpful for development in case the database schema has changed and is now incompatible - it allows refreshing data from the server without requiring to clear the entire app data.