feat(setup): improve user presence logic
Given the previous work to clear the database on a fresh app launch, it was possible to have an empty database but be marked as "signed in", which led to an empty dashboard screen. The logged in logic now takes into account whether a `me` entity is present in the database, which truly marks a logged in user.