fix(network): use interceptor to set the Monica API
The previous strategy of recreating the Retrofit instance was a very hacky workaround to make sure the API calls always took into account the latest URL provided by the user.

The better approach is to use an interceptor that sets the Monica API URL to whatever the latest value set is.