feat(contacts): support editing gender
This was a journey because I wanted to add the gender code into a separate module. To make it work neatly, I also needed to extract the `Dispatcher` API into a separate module. This of course, meant I needed to make a pure JVM plugin to apply to my new gender module. That then came with needing to add the core Hilt package as well as the core coroutines package.