feat(app): sync using WorkManager
There was a crash on a fresh install caused by the gender repository attempting to refresh without there being an access token. This made it clear that that refresh strategy didn't work.

Instead, this commit moves the solution closer to what's in NowInAndroid and creates a SyncWorker that is mainly responsible for refreshing the critical data either once at app startup or after logging in.

Currently, the gender repository is considered "critical data" but this will change into a different strategy, perhaps marking repositories as stale instead...