app: fix frequent syncing
There was no state in the app that kept track of whether any sync was done for each repository. This meant that syncing with the backend would happen a lot more than needed, eg. when backing out and re-entering a screen, after coming back to a screen after a 5 second delay, etc...

State is now kept in a companion object for each Synchronizer. There is probably a better way of doing this without keeping state in a singleton but that's a problem for later.

Also removed the unused `HomeNavigationManager`.