feat(avatar): darken avatar backgrounds in dark mode
The avatar colours coming down from Monica can be very bright and contrast too much with the screen's dark background. Additionally, the initials are light-coloured so the background should be dark enough for the text to be readable.

These changes apply the `darken` function from material-kolor in order to get a darker avatar background.