feat: replace dagger / hilt with kotlin-inject
Hilt is very useful in cutting down boilerplate so a simple migration would have made the project lose that feature. Therefore a KSP plugin was created that effectively mirrors what Hilt was doing, and in particular HiltViewModel. HiltWorker has not been migrated yet as the complexity was not worth it given there's only one Worker in the project at the moment.