fix(activities): avoid duplicate syncing
Fixes duplicate sync when spending at least 5 seconds on the contact activity create/edit screen then submitting or deleting the activity. This was because the synchronizer was invoked both when submitting but also when navigating back to the activities list. This is due to the refresh triggered in `onStart` which is only meant to happen on the first load of the screen.

For now the solution is to keep track of it through static variables.