fix pagination
When the app migrated to the multiplatform paging library, a bug was introduced.

Because the way the data (journal entries, contacts, etc.) was loaded in Compose, without using the Paging library `itemKey` and `itemContentType` function, only the first 15 results were returned.

Using the right APIs fixed the problem and now all results are returned correctly.